move getting log object
