+++ mindsdb/api/mongo/utilities/__init__.py
@@ -1,0 +1,1 @@
+from .log import log    # noqa
