+++ mindsdb/api/mongo/utilities/log.py
@@ -1,0 +1,3 @@
+import logging
+
+log = logging.getLogger('mindsdb.mongodb')
