+++ mindsdb/api/mysql/mysql_proxy/data_types/mysql_datum.py
@@ -14,7 +14,7 @@ import logging
 import math
 import struct
 from mindsdb.api.mysql.mysql_proxy.libs.constants.mysql import ONE_BYTE_ENC, TWO_BYTE_ENC, THREE_BYTE_ENC, NULL_VALUE, DEFAULT_CAPABILITIES
-from mindsdb.api.mysql.mysql_proxy.controllers.log import log
+from mindsdb.api.mysql.mysql_proxy.utilities import log
 
 
 class Datum():
