+++ mindsdb/api/mysql/mysql_proxy/utilities/log.py
